windows_workgroup: Coerce the provided reboot property and add more tests
Instead of converting the passed in reboot property when we use it coerce it when it's passed in. This allows us to easily test the conversion and it makes it so the docs don't include the legacy values.

I also added moved command generation into a method and added tests for that.

Signed-off-by: Tim Smith <tsmith@chef.io>